refactor(workspace): extract useReextraction hook
Move re-extraction state, pending-rerun banners, and confirm flow into a hook.
