+++ frontend/src/pages/Workspace/index.tsx
@@ -47,11 +47,8 @@ import MissingDocumentsSection from '@/components/SchemaEditor/MissingDocumentsS
 import TableFeedbackWidget from '@/components/TableFeedbackWidget/TableFeedbackWidget';
 import {
   WS_DISCONNECTED_REFRESH_INTERVAL,
-  getAvailableProviders,
-  getDefaultModelForProvider,
-  type LLMProviderKey,
 } from '@/constants';
-import api, { configAPI, loadAPI, schemaAPI, schematiqAPI, unitsAPI } from '@/services/api';
+import api, { configAPI, loadAPI, schematiqAPI, unitsAPI } from '@/services/api';
 import webSocketService from '@/services/websocket';
 import type {
   CostEstimate,
@@ -61,7 +58,6 @@ import type {
   ReextractionCompletedData,
   ReextractionFailedData,
   ReextractionProgressData,
-  ReextractionRequest,
   ReextractionStartedData,
   ScheMatiQConfig,
   ScheMatiQStatus,
@@ -70,7 +66,6 @@ import type {
   WebSocketMessage,
 } from '@/types';
 import type { DocumentListResponse } from '@/types/unit';
-import { getApiKeyForProvider, getConfiguredProviders } from '@/utils/apiKeyStorage';
 
 import { ChatPanel } from './chat/ChatPanel';
 import { emptyData, SHEETS, cellFormatKey } from './constants';
@@ -88,15 +83,14 @@ import { PendingRerunBanner } from './PendingRerunBanner';
 import { ProjectDetailsDialog } from './ProjectDetailsDialog';
 import { SpreadsheetChrome } from './SpreadsheetChrome';
 import { useAddDocuments } from './hooks/useAddDocuments';
+import { useReextraction } from './hooks/useReextraction';
 import { useWorkspaceLayout } from './hooks/useWorkspaceLayout';
 import { SpreadsheetSurface } from './SpreadsheetSurface';
 import type {
   CellFormatMap,
-  PendingRerunKind,
   SheetId,
   SheetSelection,
   TableDisplayOptions,
-  WorkspaceReextractionState,
   WorkspaceSessionMode,
 } from './types';
 
@@ -125,9 +119,6 @@ function Workspace() {
   const [sessionMode, setSessionMode] = useState<WorkspaceSessionMode>(requestedMode);
   const [projectDialogOpen, setProjectDialogOpen] = useState(!sessionId);
   const [detailsDialogOpen, setDetailsDialogOpen] = useState(false);
-  const [pendingRerunKind, setPendingRerunKind] = useState<PendingRerunKind | null>(null);
-  const [pendingSchemaColumns, setPendingSchemaColumns] = useState<string[]>([]);
-  const [rerunStarting, setRerunStarting] = useState(false);
   const [status, setStatus] = useState<ScheMatiQStatus | null>(null);
   const [session, setSession] = useState<VisualizationSession | null>(null);
   const [schema, setSchema] = useState<SchemaData | null>(null);
@@ -190,12 +181,7 @@ function Workspace() {
     isDraggingDivider,
     startDividerDrag,
   } = useWorkspaceLayout();
-  const [reextraction, setReextraction] = useState<WorkspaceReextractionState | null>(null);
   const [wsConnected, setWsConnected] = useState(false);
-  const [reextractConfirm, setReextractConfirm] = useState<{ columns: string[] } | null>(null);
-  const [reextractAvailability, setReextractAvailability] = useState<DocumentAvailabilityResponse | null>(null);
-  const [reextractAvailabilityLoading, setReextractAvailabilityLoading] = useState(false);
-  const [stoppingReextraction, setStoppingReextraction] = useState(false);
 
   const deferredDataRef = useRef<PaginatedData | null>(null);
   const cancelChatPendingRef = useRef<(() => Promise<boolean>) | null>(null);
@@ -341,6 +327,36 @@ function Workspace() {
     toast,
   });
 
+  const {
+    pendingRerunKind,
+    pendingSchemaColumns,
+    rerunStarting,
+    reextraction,
+    setReextraction,
+    reextractConfirm,
+    setReextractConfirm,
+    reextractAvailability,
+    setReextractAvailability,
+    reextractAvailabilityLoading,
+    stoppingReextraction,
+    clearPendingRerun,
+    runReextractPrecheck,
+    requestReextraction,
+    confirmReextraction,
+    stopReextraction,
+    startSchemaRediscovery,
+    notifyEditFollowUp,
+    runPendingEdits,
+  } = useReextraction({
+    sessionId,
+    sessionMode,
+    schema,
+    refresh,
+    setActiveSheet,
+    cancelChatPendingIfAny,
+    toast,
+  });
+
   useEffect(() => {
     setProjectDialogOpen(!sessionId);
     setSessionMode(requestedMode);
@@ -365,10 +381,6 @@ function Workspace() {
   }, []);
 
   useEffect(() => {
-    setPendingRerunKind(null);
-    setPendingSchemaColumns([]);
-    setRerunStarting(false);
-    setReextraction(null);
     setWsConnected(false);
     // Reset session-scoped view/data so a freshly imported/loaded session
     // never renders the previous session's sheet, schema, or rows.
@@ -757,259 +769,6 @@ function Workspace() {
       if (importInputRef.current) importInputRef.current.value = '';
     }
   }, [navigate, toast]);
-
-  const clearPendingRerun = useCallback(() => {
-    setPendingRerunKind(null);
-    setPendingSchemaColumns([]);
-  }, []);
-
-  const markRerunNeeded = useCallback((kind: PendingRerunKind, columns: string[] = []) => {
-    if (kind === 'unit') {
-      setPendingRerunKind('unit');
-      setPendingSchemaColumns([]);
-      return;
-    }
-
-    setPendingRerunKind((current) => current === 'unit' ? 'unit' : 'schema');
-    setPendingSchemaColumns((current) => {
-      const merged = new Set(current);
-      columns
-        .map((column) => column.trim())
-        .filter(Boolean)
-        .forEach((column) => merged.add(column));
-      return Array.from(merged);
-    });
-  }, []);
-
-  // Pre-check source-document availability for the current session. Used by the
-  // re-extract confirm card so the user can upload missing originals (e.g. after
-  // loading a project from JSON, where documents are not bundled) before running
-  // the extraction model. Failures leave availability null and fall back to the
-  // backend gate in start_gated_reextraction, which still blocks and reports.
-  const runReextractPrecheck = useCallback(async () => {
-    if (!sessionId) return;
-    setReextractAvailabilityLoading(true);
-    try {
-      const availability = await schemaAPI.precheckDocuments(sessionId, {
-        operation_type: 'reextraction',
-      });
-      setReextractAvailability(availability);
-    } catch {
-      setReextractAvailability(null);
-    } finally {
-      setReextractAvailabilityLoading(false);
-    }
-  }, [sessionId]);
-
-  // Resolve the requested columns to a concrete, non-empty target set and open
-  // the confirm card. Both routes (manual button + chat tool) re-extract only
-  // after an explicit confirm; the backend gated action owns baseline capture
-  // and document precheck so this stays purely presentational.
-  const requestReextraction = useCallback((columns?: string[]) => {
-    if (!sessionId || rerunStarting) return;
-
-    const schemaColumnNames = new Set(
-      (schema?.schema || [])
-        .map((column) => column.name)
-        .filter((name): name is string => Boolean(name)),
-    );
-    const requestedColumns = (columns && columns.length > 0 ? columns : pendingSchemaColumns)
-      .map((name) => name.trim())
-      .filter((name) => Boolean(name) && !name.toLowerCase().endsWith('_excerpt'));
-    const targetColumns = (requestedColumns.length > 0 ? requestedColumns : Array.from(schemaColumnNames))
-      .filter((name) => schemaColumnNames.has(name));
-
-    if (targetColumns.length === 0) {
-      toast({
-        title: 'No columns to re-extract',
-        description: 'Add schema columns first, then try again.',
-        variant: 'destructive',
-      });
-      return;
-    }
-
-    setReextractConfirm({ columns: targetColumns });
-    setReextractAvailability(null);
-    void runReextractPrecheck();
-  }, [pendingSchemaColumns, rerunStarting, runReextractPrecheck, schema?.schema, sessionId, toast]);
-
-  const startReextraction = useCallback(async (targetColumns: string[]) => {
-    if (!sessionId || rerunStarting || targetColumns.length === 0) return;
-
-    const chatPendingCleared = await cancelChatPendingIfAny();
-    if (!chatPendingCleared) {
-      toast({
-        title: 'Chat confirmation still pending',
-        description: 'Cancel the chat confirmation card first, then try again.',
-        variant: 'destructive',
-      });
-      return;
-    }
-
-    setRerunStarting(true);
-    try {
-      const cfg = await configAPI.getConfig().catch(() => ({ allow_llm_config: true }));
-      const configured = await getConfiguredProviders();
-      const available = getAvailableProviders(configured);
-      const provider: LLMProviderKey = !cfg.allow_llm_config
-        ? 'gemini'
-        : (available[0] ?? 'gemini');
-      const model = getDefaultModelForProvider(provider);
-      const apiKey = await getApiKeyForProvider(provider);
-      const request: ReextractionRequest = { columns: targetColumns };
-      if (apiKey) {
-        request.llm_config = { provider, model, api_key: apiKey, temperature: 0 };
-      }
-
-      const response = await schemaAPI.startReextraction(sessionId, request);
-      const docCount = response.rows_to_process || response.estimated_papers || 0;
-      if (docCount === 0) {
-        toast({
-          title: 'No source documents',
-          description: 'Upload documents or open a ScheMatiQ project with source files, then try again.',
-          variant: 'destructive',
-        });
-        return;
-      }
-
-      clearPendingRerun();
-      setActiveSheet('data');
-      setReextraction({
-        operationId: response.operation_id,
-        columns: response.columns,
-        progress: 0,
-        processedDocuments: 0,
-        totalDocuments: docCount,
-        currentColumn: response.columns[0],
-      });
-      toast({
-        title: 'Re-extraction started',
-        description: `Re-extracting ${response.columns.join(', ')} across ${docCount} document(s). Other columns stay unchanged.`,
-        duration: 4000,
-      });
-    } catch (err: any) {
-      toast({
-        title: 'Re-extraction failed to start',
-        description: err?.response?.data?.detail || err?.message || 'Could not start re-extraction',
-        variant: 'destructive',
-      });
-    } finally {
-      setRerunStarting(false);
-    }
-  }, [cancelChatPendingIfAny, clearPendingRerun, rerunStarting, sessionId, toast]);
-
-  const confirmReextraction = useCallback(async () => {
-    if (!reextractConfirm) return;
-    // Belt-and-suspenders: the Confirm button is disabled when no documents are
-    // available, but guard here too so a stale click can't bypass the gate.
-    if (reextractAvailability && !reextractAvailability.can_proceed) return;
-    const { columns } = reextractConfirm;
-    setReextractConfirm(null);
-    setReextractAvailability(null);
-    await startReextraction(columns);
-  }, [reextractAvailability, reextractConfirm, startReextraction]);
-
-  // Cancel a running re-extraction. Reuses the same backend stop mechanism as
-  // the classic Visualizer (POST /schema/stop-reextraction); the WebSocket
-  // 'reextraction_stopped' handler clears the spinner.
-  const stopReextraction = useCallback(async () => {
-    if (!sessionId || !reextraction?.operationId || stoppingReextraction) return;
-    setStoppingReextraction(true);
-    try {
-      await schemaAPI.stopReextraction(sessionId, reextraction.operationId);
-      toast({
-        title: 'Stopping re-extraction',
-        description: 'Finishing the current document, then stopping. Partial results are kept.',
-      });
-    } catch (err: any) {
-      toast({
-        title: 'Could not stop re-extraction',
-        description: err?.response?.data?.detail || err?.message || 'Stop request failed.',
-        variant: 'destructive',
-      });
-      setStoppingReextraction(false);
-    }
-  }, [reextraction?.operationId, sessionId, stoppingReextraction, toast]);
-
-  useEffect(() => {
-    if (!reextraction) setStoppingReextraction(false);
-  }, [reextraction]);
-
-  const startSchemaRediscovery = useCallback(async () => {
-    if (!sessionId || rerunStarting) return;
-
-    if (sessionMode !== 'schematiq') {
-      toast({
-        title: 'Rediscovery needs a ScheMatiQ run',
-        description: 'Imported static projects can edit the observation unit, but rediscovering schema requires a ScheMatiQ project with source documents.',
-        variant: 'destructive',
-      });
-      return;
-    }
-
-    setRerunStarting(true);
-    try {
-      const chatPendingCleared = await cancelChatPendingIfAny();
-      if (!chatPendingCleared) {
-        toast({
-          title: 'Chat confirmation still pending',
-          description: 'Cancel the chat confirmation card first, then try again.',
-          variant: 'destructive',
-        });
-        return;
-      }
-      await schematiqAPI.resume(sessionId);
-      clearPendingRerun();
-      toast({
-        title: 'Schema rediscovery started',
-        description: 'Rediscovering schema from the updated observation unit.',
-        duration: 4000,
-      });
-      await refresh({ silent: true });
-    } catch (err: any) {
-      toast({
-        title: 'Schema rediscovery failed',
-        description: err?.response?.data?.detail || err?.message || 'Could not start schema rediscovery',
-        variant: 'destructive',
-      });
-    } finally {
-      setRerunStarting(false);
-    }
-  }, [cancelChatPendingIfAny, clearPendingRerun, refresh, rerunStarting, sessionId, sessionMode, toast]);
-
-  const notifyEditFollowUp = useCallback((kind: PendingRerunKind, columns: string[] = []) => {
-    markRerunNeeded(kind, columns);
-
-    if (kind === 'unit') {
-      // The persistent top banner (driven by markRerunNeeded above) already
-      // surfaces the "Rediscover schema & re-extract" action, so we do not fire a
-      // competing toast. Exception: in an imported (non-ScheMatiQ) project
-      // rediscovery is impossible and the banner's action only errors, so we
-      // surface a proactive, action-less explanation instead.
-      if (sessionMode !== 'schematiq') {
-        toast({
-          title: 'Observation unit updated',
-          description: 'Imported static projects can edit the unit, but rediscovery needs a ScheMatiQ project with source documents.',
-        });
-      }
-      return;
-    }
-
-    // Schema edits: the persistent "Schema changed" banner is the single,
-    // always-fresh entry point for re-extraction, so we raise no duplicate toast
-    // here. The old toast was also broken — its action captured a stale
-    // `requestReextraction` closure (schema not yet refreshed when the toast was
-    // created), producing a false "No columns to re-extract" error on click.
-  }, [markRerunNeeded, sessionMode, toast]);
-
-  const runPendingEdits = useCallback(async () => {
-    if (!sessionId || !pendingRerunKind || rerunStarting) return;
-    if (pendingRerunKind === 'unit') {
-      await startSchemaRediscovery();
-      return;
-    }
-    requestReextraction(pendingSchemaColumns);
-  }, [pendingRerunKind, pendingSchemaColumns, rerunStarting, requestReextraction, sessionId, startSchemaRediscovery]);
 
   const progressPercent = Math.round((status?.progress || 0) * 100);
   const topbarQuestion = schema?.query || config?.query || '';
